move managed transform
