+++ mol_gnn/transforms/__init__.py
@@ -1,4 +1,5 @@
-from .base import Transform, ManagedTransform, Pipeline
+from .managed import ManagedTransform
+from .base import Transform, Pipeline
 from .graph import MolToGraph
 from .chem import SmiToMol
 from .mol import MolToFP
